requirements-freeze.txt
a2wsgi==1.10.10
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asgiref==3.10.0
asttokens==3.0.0
astunparse==1.6.3
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
bentoml==1.4.28
billiard==4.2.2
cachetools==6.2.1
cattrs==23.1.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-option-group==0.5.9
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.10
dvc==3.63.0
dvc-data==3.16.12
dvc-gs==3.0.2
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.2.1
filelock==3.20.0
flatbuffers==25.9.23
flatten-dict==0.4.2
flufl-lock==8.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gast==0.6.0
gcsfs==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grandalf==0.8
grpcio==1.76.0
gto==1.9.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.2
hydra-core==1.3.2
idna==3.11
importlib-metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
iterative-telemetry==0.0.10
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jupyter-client==8.6.3
jupyter-core==5.9.1
kagglehub==0.3.13
kantoku==0.18.3
keras==3.12.0
kiwisolver==1.4.9
kombu==5.5.4
libclang==18.1.1
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.5.3
multidict==6.7.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.17.0.29
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.9.86
oauthlib==3.3.1
omegaconf==2.3.0
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-aiohttp-client==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
opt-einsum==3.4.0
optree==0.17.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
pip-requirements-parser==32.0.1
platformdirs==4.5.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydot==4.0.1
pygit2==1.19.0
pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rsa==4.9.1
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.14
schema==0.7.8
scikit-learn==1.7.2
scipy==1.16.3
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simple-di==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.50.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tomli-w==1.2.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.1.1
wcwidth==0.2.14
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.22.0
zc-lockfile==4.0
zipp==3.23.0
